Navigate to the start/end of rows
Adds support for Home/CtrlA for navigating to the start of the row, and
End/CtrlE for navigating to the end of the row.  Unfortunately it
doesn't look like termbox makes it easy to determine if Shift is
currently pressed so Shift+Home for first row and Shift+End for last row
doesn't seem to be (easily?) possible.